fix(frontend): separate onboarding page metadata from client code
